+++ client/src/components/signup.tsx
@@ -6,7 +6,7 @@ import {Button, Col, Container, Form, Row} from 'react-bootstrap';
 // Custom
 import {useAuth} from '../context/authContext';
 import {notificationService} from '../services/notification.service';
-import {handleIncorrectInput, sendRequest} from './shared/inputValidation';
+import {handleIncorrectInput, sendRequest} from './common/inputValidation';
 
 const Signup = () => {
   const {setUser, setIsAuthenticated} = useAuth();
