+++ server/api/diskLogs.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const Disk = require('../models/disk.js')
-const { filterData, filterTimestampQuery } = require('./shared/filter')
+const { filterData, filterTimestampQuery } = require('./common/filter')
 const auth = require('../middleware/auth.js')
 
 // get Disk Logs with any attribute of the DiskLogs Model
